Initialize (a subset of) elements of the periodic system of elements (PSE)

Initialize the information about name, mass and isotopes. To reduce the number of decomposition hypotheses, subsets of elements can be created.

initializeElements(names)

.getElement(name, elements = NULL)

initializeCHNOPS()

initializeCHNOPSMgKCaFe()

initializeCHNOPSNaK()

initializePSE()

initializeCharges()

Arguments

names

Vector of element names within PSE.

name

Chemical element name (abbr.).

elements

Character vector of chemical element names.

Value

A list with the elements `name` repeated sum formula, `mass` nominal mass of molecule, `isotope` a list of isotopes. The function `initializeCharges` is special, since it allows to parse charges as shown in examples.

Details

These functions return full, pre-defined or user-defined (sub-) lists of elements.

References

For a description of the underlying IMS see citation("Rdisop"). Isotope patterns were obtained through wikipedia.org

Author

Steffen Neumann <sneumann@IPB-Halle.DE>

Examples

initializeCHNOPS()
#> [[1]]
#> [[1]]$name
#> [1] "C"
#> 
#> [[1]]$mass
#> [1] 12
#> 
#> [[1]]$isotope
#> [[1]]$isotope$mass
#> [1] 0.000000000 0.003354826 0.003241982
#> 
#> [[1]]$isotope$abundance
#> [1] 0.989 0.011 0.000
#> 
#> 
#> 
#> [[2]]
#> [[2]]$name
#> [1] "H"
#> 
#> [[2]]$mass
#> [1] 1
#> 
#> [[2]]$isotope
#> [[2]]$isotope$mass
#> [1] 0.007825035 0.014101790 0.016049270
#> 
#> [[2]]$isotope$abundance
#> [1] 0.99985 0.00015 0.00000
#> 
#> 
#> 
#> [[3]]
#> [[3]]$name
#> [1] "N"
#> 
#> [[3]]$mass
#> [1] 14
#> 
#> [[3]]$isotope
#> [[3]]$isotope$mass
#> [1] 0.003074002 0.000108970
#> 
#> [[3]]$isotope$abundance
#> [1] 0.99634 0.00366
#> 
#> 
#> 
#> [[4]]
#> [[4]]$name
#> [1] "O"
#> 
#> [[4]]$mass
#> [1] 16
#> 
#> [[4]]$isotope
#> [[4]]$isotope$mass
#> [1] -0.00508537 -0.00086880 -0.00083970
#> 
#> [[4]]$isotope$abundance
#> [1] 0.99762 0.00038 0.00200
#> 
#> 
#> 
#> [[5]]
#> [[5]]$name
#> [1] "P"
#> 
#> [[5]]$mass
#> [1] 31
#> 
#> [[5]]$isotope
#> [[5]]$isotope$mass
#> [1] -0.026238
#> 
#> [[5]]$isotope$abundance
#> [1] 1
#> 
#> 
#> 
#> [[6]]
#> [[6]]$name
#> [1] "S"
#> 
#> [[6]]$mass
#> [1] 32
#> 
#> [[6]]$isotope
#> [[6]]$isotope$mass
#> [1] -0.02792930 -0.02854146 -0.03213335  0.00000000 -0.03291938
#> 
#> [[6]]$isotope$abundance
#> [1] 0.9502 0.0075 0.0421 0.0000 0.0002
#> 
#> 
#> 
getMolecule("H3O+", elements=c(initializeCHNOPS(),initializeCharges()))
#> $formula
#> [1] "H3O"
#> 
#> $score
#> [1] 1
#> 
#> $exactmass
#> [1] 19.01839
#> 
#> $charge
#> [1] 0
#> 
#> $parity
#> [1] "o"
#> 
#> $valid
#> [1] "Invalid"
#> 
#> $DBE
#> [1] -0.5
#> 
#> $isotopes
#> $isotopes[[1]]
#>            [,1]         [,2]         [,3]         [,4]         [,5]
#> [1,] 19.0183897 2.002372e+01 21.022636219 2.202891e+01 2.303519e+01
#> [2,]  0.9971711 8.286234e-04  0.001999338 8.997590e-07 1.349810e-10
#>              [,6] [,7] [,8] [,9] [,10]
#> [1,] 2.404147e+01   25   26   27    28
#> [2,] 6.750000e-15    0    0    0     0
#> 
#>